+++ server/server_processor.cpp
@@ -72,7 +72,7 @@ void nibaserver::server_processor::process(nibashared::message_login &req) {
 }
 
 void nibaserver::server_processor::process(nibashared::message_fight &req) {
-    auto [self_fightable, enemy_fightable] = nibashared::prep_fight<nibashared::server_staticdata>(5, 7);
+    auto [self_fightable, enemy_fightable] = nibashared::prep_fight<nibaserver::server_staticdata>(5, 7);
     // session_.charid, req.enemyid
     nibashared::fight fight(std::move(self_fightable), std::move(enemy_fightable));
     nibashared::rng_server rng;
