+++ shared/fight.cpp
@@ -213,42 +213,4 @@ battlestats stats_computer(const attributes &attr) {
     return stats; // RVO
 }
 
-template<typename staticdata>
-fightable setup_fightable(int id) {
-    auto raw_character = staticdata::get().character(id);
-    fightable fight_character;
-    fight_character.char_data = raw_character;
-    fight_character.char_data.stats += stats_computer(raw_character.attrs);
-    for (auto magic_id : raw_character.active_magic) {
-        auto magic = staticdata::get().magic(magic_id);
-        CPRINT(raw_character.name << " has magic " << magic.name);
-        fight_character.char_data.stats += magic.stats;
-        // NOTE, aggregation initialization is copy... well fine
-        // also push_back is complaining designated initialization for reasons I don't understand
-        fight_character.magics.emplace_back(magic_ex{.cd = 0, .real_magic = magic});
-    }
-    for (auto equip_id : raw_character.equipments) {
-        auto equipment = staticdata::get().equipment(equip_id);
-        CPRINT(raw_character.name << " has item " << equipment.name);
-        fight_character.char_data.stats += equipment.stats;
-    }
-    return fight_character; // RVO
-}
-
-template fightable setup_fightable<nibashared::client_staticdata>(int id);
-template fightable setup_fightable<nibashared::server_staticdata>(int id);
-
-template<typename staticdata>
-std::pair<std::vector<fightable>, std::vector<fightable>> prep_fight(int id_me, int id_you) {
-    CPRINT("prep " << id_me << " " << id_you);
-    // initializer list do not like moves
-    std::pair<std::vector<fightable>, std::vector<fightable>> ret;
-    ret.first.push_back(setup_fightable<staticdata>(id_me));
-    ret.second.push_back(setup_fightable<staticdata>(id_you));
-    return ret; // RVO
-}
-
-template std::pair<std::vector<fightable>, std::vector<fightable>> prep_fight<nibashared::client_staticdata>(int id_me, int id_you);
-template std::pair<std::vector<fightable>, std::vector<fightable>> prep_fight<nibashared::server_staticdata>(int id_me, int id_you);
-
 } // namespace nibashared
